+++ components/tasks/TaskListItem.tsx
@@ -23,37 +23,8 @@ import {
   MessageSquareIcon,
   MoreVerticalIcon
 } from 'lucide-react'
-
+import { getPriorityColor, getStatusColor, getEffortColor } from '@/lib/colors/colors'
 const TaskListItem = React.memo(({ task, isSelected, onClick }: { task: Task; isSelected: boolean; onClick: () => void }) => {
-  const getEffortColor = (effort: string) => {
-    switch (effort) {
-      case 'backend': return 'bg-purple-200 text-purple-800'
-      case 'frontend': return 'bg-pink-200 text-pink-800'
-      case 'backend + frontend': return 'bg-indigo-200 text-indigo-800'
-      default: return 'bg-gray-200 text-gray-800'
-    }
-  }
-
-  const getPriorityColor = (priority: string) => {
-    switch (priority) {
-      case 'low': return 'border-l-green-500'
-      case 'medium': return 'border-l-yellow-500'
-      case 'high': return 'border-l-red-500'
-      case 'urgent': return 'border-l-red-500'
-      case 'critical': return 'border-l-red-500'
-      default: return 'border-l-gray-400'
-    }
-  }
-
-  const getStatusColor = (status: string) => {
-    switch (status) {
-      case 'todo': return 'bg-gray-100 text-gray-800'
-      case 'inProgress': return 'bg-blue-100 text-blue-800'
-      case 'done': return 'bg-purple-100 text-purple-800'
-      default: return 'bg-gray-100 text-gray-800'
-    }
-  }
-
   return (
     <Card 
       className={`mb-4 cursor-pointer hover:shadow-md transition-shadow duration-200 ${isSelected ? 'ring-2 ring-primary' : ''}`}
